recode the brightness function @hebing  * recode function of brightness by gsettings

+++ plugins/system/display/widget.h
@@ -123,7 +123,7 @@ class Widget : public QWidget
     //是否禁用屏幕
     void checkOutputScreen(bool judge);
     //设置屏幕亮度
-    void setBrightnessScreen(float brightnessValue);
+    void setBrightnessScreen(int brightnessValue);
     //设置亮度滑块数值
     void setBrightnesSldierValue(QString screeName);
     //保存屏幕亮度配置
